backend/auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

pydantic==2.8.2
pydantic[email]==2.8.2

SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2

bcrypt==4.3.0
passlib[bcrypt]==1.7.4

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
